Fix clamping of lookup values in rview.
Previously, the interpolators clamped the values to the intensity range of the input image. This is, however, only required by rview due to the use of the lookup table which maps intensities to color values. By clamping the values upon each lookup, there is no need to modify the interpolators to work as expected by rview alone. Instead, rview takes care of the proper handling of these boundary cases.